@@ -1,26 +1,23 @@
 package com.yarolegovich.discretescrollview.context;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.test.espresso.IdlingResource;
-import android.support.test.espresso.idling.CountingIdlingResource;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.Gravity;
 import android.view.ViewGroup;
 import android.widget.FrameLayout;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.test.espresso.IdlingResource;
+import androidx.test.espresso.idling.CountingIdlingResource;
+
 import com.yarolegovich.discretescrollview.DiscreteScrollView;
 import com.yarolegovich.discretescrollview.R;
-import com.yarolegovich.discretescrollview.transform.ScaleTransformer;
 
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
 import static android.view.ViewGroup.LayoutParams.MATCH_PARENT;
 import static android.view.ViewGroup.LayoutParams.WRAP_CONTENT;
@@ -29,7 +26,7 @@ import static android.view.ViewGroup.LayoutParams.WRAP_CONTENT;
  * Created by yarolegovich on 2/4/18.
  */
 
-public class TestActivity extends AppCompatActivity implements DiscreteScrollView.ScrollStateChangeListener {
+public class TestActivity extends AppCompatActivity implements DiscreteScrollView.ScrollStateChangeListener<RecyclerView.ViewHolder> {
 
     private DiscreteScrollView scrollView;
     private TestAdapter adapter;
